third_party/blink/web_tests/platform/android/external/wpt/pointerevents/pointerlock/pointerevent_movementxy_with_pointerlock-expected.txt

This is a testharness.js-based test.
[FAIL] mouse pointerevent movementX/Y with pointerlock test
  assert_equals: expected 5 but got 4.4554595947265625
Harness: the test ran to completion.
